fix(ci): integration test PG password + redundant lint directives
Two fixes lifted from later phases (phase-5 needed both).

1. Extract `build_pg_listen_connection` to spec/integration_helper.rb
   with the password parsed from PGBUS_DATABASE_URL. Each streams
   integration spec previously defined its own copy of this helper
   and every copy forgot to forward the password — fine on a local
   trust-auth Postgres, but every CI run failed with `fe_sendauth: no
   password supplied` because the workflow uses authenticated
   `postgres://pgbus:pgbus@...` URLs. The shared helper kills the
   duplication and the bug at the same time.

   Removes the local copies in page_born_stale_spec.rb and
   reconnect_spec.rb (the only two streams integration specs on
   phase-5; later phases will get the same removal as they merge up
   the stack).

2. spec/integration/signal_handling_spec.rb — drop three redundant
   `# rubocop:disable` directives. The .rubocop.yml exclusions for
   `spec/integration/**` cover these cops, so the directives have
   been linting as `RedundantCopDisableDirective` since phase-5.
